Title changes with color submission
"Colorizer" title now reads "'Colorizer:  this.state.color" when a color is submitted with the 'go' button (ex: type in 'red' and hit go - now the square is red and the title is: 'Colorizer: red'); This was supposed to be done via a portal, but I could not get the portal to target something that was recognized as a DOM element so I could not get it to work. The portal code is commented out to show there was an attempt and to reference if I come back to try again in the future.